==> 02_Executive_Summary.Rmd <==
---
output: word_document
---
# Figures
Page 

# Tables
Page

# Appendices
Page

# Executive Summary

Channel Islands National Park (CINP) has conducted long-term ecological monitoring of the kelp forests around San Miguel, Santa Rosa, Santa Cruz, Anacapa and Santa Barbara Islands since 1982. The original permanent transects were established at 16 sites between 1981 and 1986 with the first sampling beginning in 1982, 2019 being the 38th year of monitoring. An additional site, Miracle Mile, was established at San Miguel Island in 2001 by a commercial fisherman with assistance from the park. Miracle Mile was partially monitored from 2002-2004, and then fully monitored (using all Kelp Forest Monitoring Program (KFMP) protocols) since 2005. Only two original sites were inside a Marine Protected Area (MPA) since 1982 and it become apparent that the sample size inside and outside MPAs were too small for statistical comparisons. This type of comparison was deemed important by park management as well as other State and Federal agencies. In 2005, 16 additional permanent sites were established to collect baseline data from inside and adjacent to four State Marine Reserves (SMRs) that were established in 2003. these new sites were chosen to compliment existing sites and to best study the differences in reserve status by island. Twenty four of the 33 KFMP sites now constitute our SMR reference sites. This report thoroughly examines the effects of SMRs.

The 2005-2019 monitoring efforts utilized `r Vessel_Time` days of vessel time to conduct `r Total_Dives` dives for a total of `r Dive_Time` hours (`r round(Dive_Time/24, 0)` days) of bottom time at 33 monitoring sites. An average of `r round(Divers, 0)` divers participate in KFMP each year. Population dynamics of a select list of 71 “indicator species” (consisting of taxa or categories of algae, fish, and invertebrates) were measured at the 33 permanent sites. In addition, population dynamics were measured for all additional species of fish observed at the sites during the roving diver fish count. Survey techniques follow the CINP Kelp Forest Monitoring Protocol Handbooks (Davis et al., 1997, and Kushner and Sprague, I/P). The techniques utilize SCUBA and surface-supplied-air to conduct the following monitoring protocols: 1 m² quadrats, 5 m² quadrats, band transects, random point contacts, fish transects, roving diver fish counts, video transects, size frequency measurements, and artificial recruitment modules. Hourly temperature data were collected using remote temperature loggers at 32 sites, the exception being Miracle Mile where there is no temperature logger installed. This community trend report contains a summary of survey methods used, statistical methods used for analysis, results of analysis, and a discussion of the results. 

The status of kelp forests... inside trends vs outside trends... SSWD... Urchin disease

2005-2019 broad oceanographic and meteorological trends... ONI, PDO, blob, temp loggers, other

SR 2005-2019 broad trends...

SC 2005-2019 broad trends...

AN 2005-2019 broad trends...

SB 2005-2019 broad trends...

# Acknowledgments 

Funding for the kelp forest monitoring program for 2005-2019 was entirely provided by the U.S. National Park Service (NPS) with most funding coming from the Stewardship of New Marine Protected Areas and some from the Inventory and Monitoring Program. The monitoring program is conducted in cooperation with the California Department of Fish and Wildlife (CDFW) and the U.S. Department of Commerce, National Oceanographic and Atmospheric Administration (NOAA), Marine Sanctuary Program.

We are deeply indebted to the many divers who have participated in this program. As of 2019, 429 divers have participated in KFMP. These 429 divers have been permanent employees, seasonal employees, interns, and volunteers. All volunteer divers were trained and/or certified with agencies such as NOAA, CDFW, Aquariums and/or Universities. Past employees have also returned to volunteer for KFMP dive trips. Without this volunteer base of well-trained and qualified divers it would be impossible to conduct this program at its current funding level. We also greatly appreciate the efforts of our boat captains and Park Dive Officers for ensuring that all our operations run safely and successfully. We would like to especially thank our incredible NPS Seasonal Kelp Forest Monitoring Biological Science Technicians.


# List of Acronyms 

| **Acronym** | **Definition** |
|:------|:-----|
| ARM | Artificial Recruitment Module |
| CDFW | California Department of Fish and Wildlife |
| CINMS | Channel Islands National Marine Sanctuary |
| CINP | Channel Islands National Park |
| CPC | Climate Prediction Center |
| ENSO | El Niño-Southern Oscillation | 
| FSC | Foliar Standing Crop | 
| KFMP | Kelp Forest Monitoring |
| MPA | Marine Protected Area Program |
| nMDS | Non-metric Multidimensional Scaling |
| NOAA | National Oceanic and Atmospheric Administration |
| NPS | National Park Service |
| NRPP | Natural Resources Preservation Program |
| PDO | Pacific Decadal Oscillation |
| PISCO | The Partnership for Interdisciplinary Studies of Coastal Oceans |
| ONI | Oceanic Niño Index  |
| RCCA | Reef Check California |
| SBC LTER | Santa Barbara Coastal Term Ecological Research |
| SST | Sea Surface Temperature |
| SSWD | Sea Star Wasting Disease |
| SMCA | State Marine Conservation Area |
| SMR | State Marine Reserve | 
| RPC | Random Point Contact |
| UCSB | University of California, Santa Barbara |
