==> 06_Discussion.Rmd <==
---
output: word_document
---
# Discussion

Diversity differences – Recent significant increase inside of MPAs, SR recently decreasing while other islands increasing

Similarity differences - SR distinct 2005-2010, all islands similar 2010-2015, SR distinct again 2016-2019

Biomass trends of species identified in ISA are consistent with other observations in CA kelp forests (kelp deforestation, urchin proliferation, sea star and abalone declines) in the previous decade

Species interactions and biomass/density trends trends…

+    Trend in sea stars and 2013-2014 SSWD 

+ pre blob/El Niño major decline, then blob = perfect storm

+ Santa Rosa Island

+    ISA show Pycnopodia as an indicator inside and outside at Santa Rosa up to 2013

+    Purple urchins identified by ISA in 2015 inside and 2016 outside and continue to be identified after.

+    Red urchins identified by ISA in 2015 inside and 2018 outside

+    Pycnopodia last predator left at Santa Rosa, when they go, there is nothing left to effectively control urchins

+ Biomass of sheep head goes up inside and out, with increase in urchins

+    Main driver for decrease in diversity at SR

+ Inside stays more diverse indicating MPA resiliency when confronted with major changes

+    Urchin biomass increases dramatically as kelp biomass decreases dramitically esp. outside reserve

+    Red abalone biomass increases initially (increased competition, more foraging) 

+ decline later when urchin biomass peaks (out competed)

+    Kelp no longer identified by ISA outside reserve post 2014, indicates dominance of urchins and impact on kelp

+ Tethya have major decline with latest warm water event

+ Could be strong correlation with PDO phases (most effected at Santa Rosa)

+ Santa Cruz/Anacapa/Santa Barbara

+ Santa Cruz/Anacapa

+ General interactions

+ Garibaldi never indicate inside-MPA communities, only outside, and have significantly higher biomass outside. Perhaps reduced competition due to fishing pressure?

+ Halfmoon and opaleye (herbiverous fish) have higher inside/outside biomass ratio where algaes tend to be more plentiful

+ reference ratio plot and RPC algal cover plots

Fisheries effects - Lobster and Warty Cucumber - with original 16 plot set showing the effects of old reserves at ANI vs new reserves. I want to use these two examples to draw out points about resiliency, edge effects and "effective MPA size", and larval connectivity

Prior to the new reserve implementation, spiny lobsters exhibited a relatively stable population that was larger inside the old reserves than outside, it’s also worth noting that these old reserves were expanded with the implementation of the new reserves, and following that we see increases across the board (including a small one outside the reserves), but the greatest increase occurs in these oldest MPAs. This is consistent with the notion that edge effects in smaller MPAs are proportionally higher, and leave a proportionally smaller inside area “truly” protected, as well as the notion that the older an MPA is, the higher the resiliency of its community. Lobster are also an important predator of urchins, whose herbivory has major impacts on kelp abundance and biomass. In the coming weeks we’ll show you some more species, including kelp, urchins, and sunflower stars (another important urchin predator), but according to our random forest analysis, spiny lobster are one of the two most important species in identifying whether a community is from inside or outside an MPA.

Warty sea cucumber are  another interesting species for the region, since it had a diving-based fishery begin in the early 1990s, although harvesting began prior to that year. We can see here that the abundance of warty sea cucumbers was relatively even between inside and outside- MPA communities, until the beginning of the fishery, where the outside (at the time) MPA sites experienced a steady decline, as the protected areas saw increases in abundance. If you stop in 2001, this is a fairly standard story about fishery effects – however, the interesting part of this species’ trend is that, rather than staying relatively consistent, as it appears to pre-fishery, the inside-MPAs experience an initial increase, followed by a decline approximately to the level of the new MPAs, while the new MPAs are beginning to show an increase in density, and the outside reserve sites continue to decline. For many marine species that have a long larval period, recruitment is not very site-specific, and larval connectivity in the metapopulation as a whole is important for recruitment and population consistency, so this trend is likely due to overall reductions in larval abundance available for recruitment, due to smaller spawning populations outside MPAs. While the new MPAs also show signs of resilience and recovery in this case, warty sea cucumbers highlight the importance of these interconnected communities inside and outside MPAs, and thus that conditions and events outside of MPAs are still capable of affecting those inside.
